moved casts to short into Schematic constructor

+++ src/com/selfdot/Component.java
@@ -2,6 +2,7 @@ package com.selfdot;
 
 public class Component {
 
+
     private final Schematic schematic;
     public Component(Schematic schematic) {
         this.schematic = schematic;
@@ -26,7 +27,7 @@ public class Component {
                 2, 0,
                 3, 0
         };
-        AND_GATE = new Component(new Schematic("andGate", (short) 2, (short) 3, (short) 2, andGate));
+        AND_GATE = new Component(new Schematic("andGate", 2, 3, 2, andGate));
 
         int[] xorGate = new int[] {
                 0, 0, 1, 0, 0,
@@ -41,7 +42,7 @@ public class Component {
                 2, 2, 0, 0, 0,
                 3, 0, 0, 2, 0,
         };
-        XOR_GATE = new Component(new Schematic("andGate", (short) 5, (short) 3, (short) 3, xorGate));
+        XOR_GATE = new Component(new Schematic("andGate", 5, 3, 3, xorGate));
     }
 
     public Schematic getSchematic() {
